+++ src/routes/attempts/_post/questions/v1/MomentQuestion.ts
@@ -1,4 +1,4 @@
-import DropdownQuestion from "./DropdownQuestion";
+import DropdownQuestion from "../DropdownQuestion";
 
 export default class MomentQuestion extends DropdownQuestion<"moment-date"> {
   constructor() {
